FEAT: Move auction selector related tests
