+++ epochStart/metachain/systemSCs_test.go
@@ -8,7 +8,6 @@ import (
 	"math"
 	"math/big"
 	"os"
-	"strings"
 	"testing"
 
 	arwenConfig "github.com/ElrondNetwork/arwen-wasm-vm/v1_4/config"
@@ -1794,89 +1793,6 @@ func TestSystemSCProcessor_ProcessSystemSmartContractStakingV4EnabledCannotPrepa
 
 	err := s.ProcessSystemSmartContract(validatorsInfo, &block.Header{})
 	require.Equal(t, errProcessStakingData, err)
-}
-
-func TestSystemSCProcessor_ProcessSystemSmartContractStakingV4EnabledErrSortingAuctionList(t *testing.T) {
-	t.Parallel()
-
-	args, _ := createFullArgumentsForSystemSCProcessing(0, createMemUnit())
-
-	errGetNodeTopUp := errors.New("error getting top up per node")
-	nodesConfigProvider, _ := notifier.NewNodesConfigProvider(args.EpochNotifier, []config.MaxNodesChangeConfig{{MaxNumNodes: 1}})
-	argsAuctionListSelector := AuctionListSelectorArgs{
-		ShardCoordinator: args.ShardCoordinator,
-		StakingDataProvider: &mock.StakingDataProviderStub{
-			GetNodeStakedTopUpCalled: func(blsKey []byte) (*big.Int, error) {
-				switch string(blsKey) {
-				case "pubKey0", "pubKey1":
-					return nil, errGetNodeTopUp
-				default:
-					require.Fail(t, "should not call this func with other params")
-					return nil, nil
-				}
-			},
-		},
-		MaxNodesChangeConfigProvider: nodesConfigProvider,
-	}
-	als, _ := NewAuctionListSelector(argsAuctionListSelector)
-	args.AuctionListSelector = als
-
-	owner := []byte("owner")
-	ownerStakedKeys := [][]byte{[]byte("pubKey0"), []byte("pubKey1")}
-	stakingcommon.RegisterValidatorKeys(args.UserAccountsDB, owner, owner, ownerStakedKeys, big.NewInt(2000), args.Marshalizer)
-
-	validatorsInfo := state.NewShardValidatorsInfoMap()
-	_ = validatorsInfo.Add(createValidatorInfo(ownerStakedKeys[0], common.AuctionList, owner, 0))
-	_ = validatorsInfo.Add(createValidatorInfo(ownerStakedKeys[1], common.AuctionList, owner, 0))
-
-	s, _ := NewSystemSCProcessor(args)
-	args.EpochNotifier.CheckEpoch(&block.Header{Epoch: args.EpochConfig.EnableEpochs.StakingV4EnableEpoch})
-
-	err := s.ProcessSystemSmartContract(validatorsInfo, &block.Header{PrevRandSeed: []byte("rnd")})
-	require.Error(t, err)
-	require.True(t, strings.Contains(err.Error(), errGetNodeTopUp.Error()))
-	require.True(t, strings.Contains(err.Error(), epochStart.ErrSortAuctionList.Error()))
-}
-
-func TestSystemSCProcessor_ProcessSystemSmartContractStakingV4NotEnoughSlotsForAuctionNodes(t *testing.T) {
-	t.Parallel()
-
-	args, _ := createFullArgumentsForSystemSCProcessing(0, createMemUnit())
-	nodesConfigProvider, _ := notifier.NewNodesConfigProvider(args.EpochNotifier, nil)
-	argsAuctionListSelector := AuctionListSelectorArgs{
-		ShardCoordinator:             args.ShardCoordinator,
-		StakingDataProvider:          args.StakingDataProvider,
-		MaxNodesChangeConfigProvider: nodesConfigProvider,
-	}
-	als, _ := NewAuctionListSelector(argsAuctionListSelector)
-	args.AuctionListSelector = als
-
-	owner1 := []byte("owner1")
-	owner2 := []byte("owner2")
-
-	owner1StakedKeys := [][]byte{[]byte("pubKey0")}
-	owner2StakedKeys := [][]byte{[]byte("pubKey1")}
-
-	stakingcommon.RegisterValidatorKeys(args.UserAccountsDB, owner1, owner1, owner1StakedKeys, big.NewInt(2000), args.Marshalizer)
-	stakingcommon.RegisterValidatorKeys(args.UserAccountsDB, owner2, owner2, owner2StakedKeys, big.NewInt(2000), args.Marshalizer)
-
-	validatorsInfo := state.NewShardValidatorsInfoMap()
-
-	_ = validatorsInfo.Add(createValidatorInfo(owner1StakedKeys[0], common.EligibleList, owner1, 0))
-	_ = validatorsInfo.Add(createValidatorInfo(owner2StakedKeys[0], common.AuctionList, owner2, 0))
-
-	s, _ := NewSystemSCProcessor(args)
-	s.EpochConfirmed(args.EpochConfig.EnableEpochs.StakingV4EnableEpoch, 0)
-	err := s.ProcessSystemSmartContract(validatorsInfo, &block.Header{PrevRandSeed: []byte("rnd")})
-	require.Nil(t, err)
-
-	expectedValidatorsInfo := map[uint32][]state.ValidatorInfoHandler{
-		0: {
-			createValidatorInfo(owner1StakedKeys[0], common.EligibleList, owner1, 0),
-			createValidatorInfo(owner2StakedKeys[0], common.AuctionList, owner2, 0),
-		},
-	}
-	require.Equal(t, expectedValidatorsInfo, validatorsInfo.GetShardValidatorsInfoMap())
 }
 
 func TestSystemSCProcessor_ProcessSystemSmartContractStakingV4Enabled(t *testing.T) {
